Modify prototypes.
I am not sure if this is a good idea. It seems to me that there is a
choice between modifying prototypes and `for each` - you only get one or
the other because they'll screw each other up. But I think that
something like

arr.foreach(function(obj) {
    //do stuff with obj
});

is nicer anyway.

Come to think of it, it's going to mess with other people's code too (if
they use for each). Ughhhhhhhh.