Start refactor of BasicHooks.
It's actually not very helpful to have these all be callback functions.

Use case: Say you want a monster to patrol. Then it makes sense to pass
in a movement Vec into PlatformerLike, which may change over time. The
easier way to do this would be to just call PlatformerLike directly,
rather than using it like a callback.