Remove IEqual. Add permanence. Fix tricky hide() bug.
* Permanence:

If something is permanent, it'll remain on screen even after you leave
the screen and come back. If it is not permanent, it will disappear.

* IEqual

This interface was not helping me at all.

* Tricky hide() bug:

Fathom.update() maintains a list of items that need to be updated.  If
an item was hidden() during Fathom.update(), it would not be removed
from that list. This would cause update() to be called once more than it
should.