super() first thing in Text.
There was this really gnarly series of dependencies that needed to be
satisfied to call super() in Text, which made editing that file a mess.

Basically the problem was that setting the width of Text would set the
width of the textField, so you couldn't call super before the textField
existed, because that would implicitly set the width.

But you couldn't addChild(textField) because you can't addChild()
anything without calling super(). Otherwise you don't even have a
Sprite, so it's not like this is unreasonable or anything..

Fixed the problem by making super() only optionally take a width and
height. super() should always be the first call in a constructor, I've
decided. (Someone *may* have decided this before me.)