Improve camera.
This is one of those happy moments when removing code improves the
code base.

The problem was that the camera would ease it's x and y position into
their correct values. In fact, the camera should ease into it's x and y
position that have not been constrained to within the deadzone. This is
because if you're right on the edge of a map and suddenly look off it,
where you're trying to look is way off the map (the unconstrained
coordinates, not right at the edge (the contrained coordinates).

So all I had to do was remove the math to contrain the coordinates
before easing.